+++ custom_components/kirkhill_wind/config_flow.py
@@ -21,7 +21,6 @@ from .const import (
     CONF_BASE_URL,
     CONF_CREATE_DASHBOARD,
     CONF_ENABLE_PAYMENT_TRACKING,
-    CONF_GRAPH_HOURS,
     CONF_OWNER_PROJECTED_ANNUAL_EARNINGS_GBP,
     CONF_SCAN_INTERVAL,
     CONF_SITE_NAME,
@@ -29,7 +28,6 @@ from .const import (
     DEFAULT_BASE_URL,
     DEFAULT_CREATE_DASHBOARD,
     DEFAULT_ENABLE_PAYMENT_TRACKING,
-    DEFAULT_GRAPH_HOURS,
     DEFAULT_OWNER_PROJECTED_ANNUAL_EARNINGS_GBP,
     DEFAULT_SCAN_INTERVAL,
     DEFAULT_SITE_NAME,
@@ -215,10 +213,6 @@ class KirkHillWindOptionsFlow(config_entries.OptionsFlow):
                             DEFAULT_ENABLE_PAYMENT_TRACKING,
                         ),
                     ): bool,
-                    vol.Required(
-                        CONF_GRAPH_HOURS,
-                        default=current.get(CONF_GRAPH_HOURS, DEFAULT_GRAPH_HOURS),
-                    ): vol.All(int, vol.Range(min=1, max=168)),
                     vol.Required(
                         CONF_OWNER_PROJECTED_ANNUAL_EARNINGS_GBP,
                         default=current.get(